requirements.txt
requests
sseclient-py
duckdb
